Module proof_gen::proof_gen

This module defines the proof generation methods corresponding to the three types of proofs the zkEVM internally handles.

Structs

ProofGenError

A custom error type to handle failure cases during proof generation.

Functions

dummy_proof

Generates a dummy proof for a dummy circuit doing nothing. This is useful for testing purposes only.

generate_agg_block_proof

Generates an aggregation block proof from two child proofs.

generate_block_proof

Generates a block proof.

generate_segment_agg_proof

Generates an aggregation proof from two child proofs.

generate_segment_proof

Generates a transaction proof from some IR data.

generate_transaction_agg_proof

Generates a transaction aggregation proof from two child proofs.

Type Aliases

ProofGenResult

A type alias for Result<T, ProofGenError>.