Struct proof_gen::prover_state::ProverStateBuilder

pub struct ProverStateBuilder { /* private fields */ }

Builder for the prover state.

Implementations

impl ProverStateBuilder

pub const fn set_arithmetic_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_byte_packing_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_cpu_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_keccak_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_keccak_sponge_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_logic_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_memory_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_memory_before_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub const fn set_memory_after_circuit_size(self, size: Range<usize>) -> Self

Specifies a range of degrees to be supported for this STARK table’s associated recursive circuits.

pub fn build(self) -> ProverState

Instantiate the prover state from the builder. Note that this is a very expensive call!

impl ProverStateBuilder

pub fn build_verifier(self) -> VerifierState

Instantiate the verifier state from the builder. Note that this is a very expensive call!