Module proof_gen::types

This module contains type aliases and custom Error definition for convenient proof generation.

Constants

EXTENSION_DEGREE

The extension degree of the field used in the proof system.

Type Aliases

AllRecursiveCircuits

A type alias for the set of preprocessed circuits necessary to generate succinct block proofs.

Config

The recursive circuit configuration to be used to shrink and aggregate proofs.

Field

The base field on which statements are being proven.

Hash

Hasher

PlonkyProofIntern

A type alias for proofs generated by the zkEVM.

VerifierData

A type alias for the verifier data necessary to verify succinct block proofs. While the prover state AllRecursiveCircuits can also verify proofs, this VerifierData is much lighter, allowing anyone to verify block proofs, regardless of the underlying hardware.