Type Alias proof_gen::types::AllRecursiveCircuits

pub type AllRecursiveCircuits = AllRecursiveCircuits<Field, Config, EXTENSION_DEGREE>;

A type alias for the set of preprocessed circuits necessary to generate succinct block proofs.

Aliased Type

struct AllRecursiveCircuits {
    pub root: RootCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>,
    pub segment_aggregation: SegmentAggregationCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>,
    pub txn_aggregation: TxnAggregationCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>,
    pub block: BlockCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>,
    pub two_to_one_block: TwoToOneBlockCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>,
    pub by_table: [RecursiveCircuitsForTable<GoldilocksField, PoseidonGoldilocksConfig, 2>; 9],
}

Fields
root: RootCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>

The EVM root circuit, which aggregates the (shrunk) per-table recursive proofs.

segment_aggregation: SegmentAggregationCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>

The segment aggregation circuit, which verifies that two segment proofs that can either be root or aggregation proofs.

txn_aggregation: TxnAggregationCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>

The transaction aggregation circuit, which verifies the aggregation of two proofs that can either be a segment aggregation representing a transaction or an aggregation of transactions.

block: BlockCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>

The block circuit, which verifies a transaction aggregation proof and an optional previous block proof.

two_to_one_block: TwoToOneBlockCircuitData<GoldilocksField, PoseidonGoldilocksConfig, 2>

The two-to-one block aggregation circuit, which verifies two unrelated block proofs.

by_table: [RecursiveCircuitsForTable<GoldilocksField, PoseidonGoldilocksConfig, 2>; 9]

Holds chains of circuits for each table and for each initial degree_bits.