Module proof_gen::verifier_state

This module defines the VerifierState, that contains the necessary data to handle succinct block proofs verification.

Structs

VerifierState

Plonky2 verifier state.

Type Aliases

VerifierStateBuilder

Builder for the verifier state. This is essentially the same as the ProverStateBuilder, in that we need to first generate the entire prover state before extracting the verifier data.