Type Alias proof_gen::verifier_state::VerifierStateBuilder

pub type VerifierStateBuilder = ProverStateBuilder;

Builder for the verifier state. This is essentially the same as the ProverStateBuilder, in that we need to first generate the entire prover state before extracting the verifier data.

Aliased Type

struct VerifierStateBuilder { /* private fields */ }

Implementations

impl VerifierStateBuilder

pub fn build_verifier(self) -> VerifierState

Instantiate the verifier state from the builder. Note that this is a very expensive call!